backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argparse==1.4.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cors==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
future==1.0.0
galileo-core==2.34.0
galileo-observe==1.18.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-pinecone==0.2.2
langchain-tests==0.3.8
langchain-text-splitters==0.3.5
langsmith==0.2.10
marshmallow==3.25.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.59.7
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
promptquality==1.0.0
propcache==0.2.1
pydantic==2.10.5
pydantic-core==2.27.2
pydantic-settings==2.7.1
pyjwt==2.10.1
pysocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.37
starlette==0.41.3
syrupy==4.8.1
tenacity==9.0.0
tiktoken==0.8.0
tldextract==5.1.3
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zope-event==5.0
zope-interface==7.2
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-openai
langchain-pinecone
pinecone-client
galileo-observe